add get env utils function
